pyproject.toml
[tool.poetry]
name = "supabase-py-expt"
version = "0.1.0"
description = ""
authors = ["v3vapp <v3v.app@gmail.com>"]
readme = "README.md"
packages = [{include = "supabase_py_expt"}]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
supabase = "^1.0.0"
python-dotenv = "^0.21.1"
faker = "^16.8.1"
faker-commerce = "^1.0.3"
storage3 = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
